+++ extendr-api/src/scalar/rcplx_full.rs
@@ -23,11 +23,15 @@ impl CanBeNA for c64 {
 /// Rcplx has a special NA value, obtained from R headers via R_NaReal.
 ///
 /// Rcplx has the same footprint as R's complex value allowing us to use it in zero copy slices.
-#[derive(Debug)]
+#[repr(C)]
 pub struct Rcplx(c64);
 
 impl Rcplx {
     gen_impl!(Rcplx, c64);
+
+    pub fn new(re: f64, im: f64) -> Self {
+        Self(c64::new(re, im))
+    }
 
     pub fn is_nan(&self) -> bool {
         self.0.is_nan()
@@ -157,5 +161,25 @@ gen_unop!(Rcplx, Neg, |lhs: c64| Some(-lhs), "Negate a Rcplx value.");
 impl PartialEq<f64> for Rcplx {
     fn eq(&self, other: &f64) -> bool {
         self.re().inner() == *other && self.im() == 0.0
+    }
+}
+
+impl std::fmt::Debug for Rcplx {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        if self.is_na() {
+            write!(f, "NA_COMPLEX")
+        } else {
+            write!(
+                f,
+                "{:?} {} {:?}i",
+                self.re(),
+                if self.im().is_sign_negative() {
+                    '-'
+                } else {
+                    '+'
+                },
+                self.im().abs()
+            )
+        }
     }
 }