+++ extendr-api/src/wrapper/complexes.rs
@@ -12,7 +12,7 @@ use std::iter::FromIterator;
 ///     assert_eq!(vec.len(), 5);
 /// }
 /// ```  
-#[derive(Debug, PartialEq, Clone)]
+#[derive(PartialEq, Clone)]
 pub struct Complexes {
     pub(crate) robj: Robj,
 }
@@ -65,6 +65,16 @@ impl DerefMut for Complexes {
         unsafe {
             let ptr = DATAPTR(self.get()) as *mut Rcplx;
             std::slice::from_raw_parts_mut(ptr, self.len())
+        }
+    }
+}
+
+impl std::fmt::Debug for Complexes {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        if self.len() == 1 {
+            write!(f, "{:?}", self.elt(0))
+        } else {
+            f.debug_list().entries(self.iter()).finish()
         }
     }
 }